Convert calendar into a namespace
Change the remaining instances into `time_point`.
The remaining public member functions are all static, so now they are namespace functions.

+++ src/savegame.cpp
@@ -72,8 +72,8 @@ void game::serialize( std::ostream &fout )
 
     json.start_object();
     // basic game state information.
-    json.member( "turn", static_cast<int>( calendar::turn ) );
-    json.member( "calendar_start", static_cast<int>( calendar::start_of_cataclysm ) );
+    json.member( "turn", calendar::turn );
+    json.member( "calendar_start", calendar::start_of_cataclysm );
     json.member( "initial_season", static_cast<int>( calendar::initial_season ) );
     json.member( "auto_travel_mode", auto_travel_mode );
     json.member( "run_mode", static_cast<int>( safe_mode ) );
